Some progress towards a true wedge independent distance
git-svn-id: file:///home/roberto/Git/svn@12450 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 1afc28e7a63b90aeb00cbea5380197a89b6b42cb

+++ trunk/xmipp/applications/programs/mpi_classify_CL3D/mpi_classify_CL3D.h
@@ -88,7 +88,7 @@ public:
     MultidimArray< std::complex<double> > Pupdate;
 
     // Update for next iteration
-    MultidimArray< int > PupdateMask;
+    MultidimArray< double > PupdateMask;
 
     // P in cylindrical coordinates around Z
     MultidimArray<double> PcylZ;
